Declutter gulpfile.js and EJS templating optimisation
