+++ lib/util.js
@@ -162,6 +162,10 @@ module.exports = {
 
         let customTemplateContent = fs.readFileSync(customTemplate, 'utf-8');
 
+        // Pass the path options fro EJS Template
+        options.path = path.join(__dirname,
+            '../app/templates/gulptasks/');
+
         if (options) {
 
             // Custom template with options
