+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@pnp/generator-spfx",
-    "version": "1.6.0-beta1",
+    "version": "1.6.0-beta2",
     "description": "This is an open-source extension for native out of the box SPFx generator to introduce additional scaffolding support for Angular Elements, VueJS, Handlebars and many more in future. This generator has a dependency on native SPFx generator from Microsoft.",
     "author": {
         "name": "Office 365 / SharePoint PnP Core Team",
